my_math apply
Now, we don't need any default libraries such a "math.h"

+++ src/rpi_lib/delay.c
@@ -24,9 +24,9 @@ volatile unsigned long long get_systime(void) {
     chi = *(volatile unsigned int *) SYST_CHI;
     clo = *(volatile unsigned int *) SYST_CLO;
   }
-  t = chi;
+  t = (unsigned long long)chi;
   t = t << 32;
-  t += clo;
+  t += (unsigned long long)clo;
   return t;
 }
